Fixes bugs that prevented writing of binary -0 floats and limited precision when writing text floats. (#54)
Note that "%20g" refers to the width of the string representation, whereas "%.20g" means that 20 digits of precision after the decimal point is required. It's unclear what the magic number should be, but this one passes all of our tests. This is probably not the best solution, but it is the least complex one (probably by far).