Matches behavior between the text and binary readers after they're created, but before their position has been advanced.
This is just a parity gap between the text and binary ion readers, which in the normal case only impacts code with too-verbose diagnostics. In the worst case, it could be used by an SDK consumer to determine what kind of reader is currently being used, breaking encapsulation. Before the change, the unit test I added to the binary reader fails. After the change, it passes (without causing any other failures).

Now, the binary and text readers will both yield `type_none` in this situation.

https://issues.amazon.com/issues/ION-1226