Fix bug resulting in read overflow of stack allocated buffer
The original ordering of the for loop's condition resulted in
the memcpy of 4 bytes being completed before the actual bounds
checking was done. During testing, due to the starting offset,
this resulted in reading two bytes beyond the `acc` buffer.
The end result would not cause any data corruption, but
triggered sanitizer checks during debug execution.