Disable sanitizer checks where known behaviors occur
When running with sanitizers enabled some of the functionality in
decNumber and a few places within ion-c itself, triggers runtime
sanitizer checks. I've disabled specific sanitizer checks on the
functions that exhibit the behaviors in order to quiet the
sanitizers, with the expectation that a sanitizer error should
cause unit test failure. I have looked at each of the instances
and verified that despite the sanitizer being correct, the code
is behaving as expected.

That said, there are certainly ways to implement the code without
changing functionality, that would not exhibit these behaviors.
It would be ideal to revisit each case and eliminate the need for
disabling the sanitizer checks.