Fix memleak when writer is closed prior to closing containers
The writer's _pending_temp_entity_pool is moved over to
_temp_entity_pool while transitioning from the symtab
intercept state, when depth is zero.

Prior to this commit if the writer was closed prior to that
transition, the data pointed to by _pending_temp_entity_pool
would be leaked. This commit adds checking during
_ion_writer_free to ensure the pending temp entity pool is
freed if needed, similar to how the temp entity pool is freed.