Fix leak of resources when unable to flush writer
Prior to this commit when a binary writer was closed,
but could not be flushed, the _ion_writer_close_helper
function would skip _ion_writer_binary_close, which
resulted in the binary writer's _value_stream being
leaked.

This commit updates the close functions for both binary
and text formats to include a BOOL to communicate if
we want to flush or not. _ion_writer_close_helper was then
updated to track if we are able to flush and provide
that info to the format writer's close, so we can ensure
all resources are cleaned up.