Add support for down-converting ion to JSON
This commit adds functionality to the ion text writer to allow for
a basic lossy conversion to JSON encoding.

An existing issue that needs some consensus is how multiple top-level
values are handled. With this commit multiple top-level items are
serialized with no delimeter. This is supported by some JSON parsers,
such the one used by the tool jq, but may be an issue with others. A
simple hacky fix can be added to place all top-level values into a list.