Replace Assertion Failure with Error for Premature Container Ends (#356)
* Replace Assertion Failure with Error for Premature Container Ends

This change replaces what is an assertion failure with a returned
error when the user attempts to step out of an Ion Container that
was not completed before the stream ended.

The issue is causing ion-python loads to hang for this case.

GH #355

* Add check for valid state in reader next as well

I realized the text reader also throws an error when the user tries
to get the next value from within the container. That seems right
to me as well, so I added a check there and then a test to prove
consistency.